pyproject.toml
[project]
name = "01-ai-engineering-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "groq",
    "openai",
    "pydantic",
    "pydantic-settings",
    "streamlit",
    "qdrant-client",
    "langsmith",
    "langchain-openai",
    "ragas",
    "rapidfuzz",
    "ipykernel",
    "instructor",
    "fastapi",
    "uvicorn",
    "httpx",
    "jsonref",
    "open_clip_torch",
    "kafka-python"
]
